refactor: rename UserDefinition to UserDocument and fix other minor build errors
