+++ apps/server/src/modules/room/helpers.ts
@@ -1,7 +1,7 @@
-import { UserDefinition } from "../user/UserModel";
+import { UserDocument } from "../user/UserModel";
 import { RoomDefinition } from "./RoomModel";
 
-export const isRoomMember = (room: RoomDefinition, user: UserDefinition) => {
+export const isRoomMember = (room: RoomDefinition, user: UserDocument) => {
   return room
     .participants
     .map(u => u._id.toString())
